chore: move nuxt config to module
